+++ src/pages/ai/AiServiceDetail.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import * as S from "./style";
 
 // 컴포넌트
@@ -8,12 +8,39 @@ import { AiServiceDetailTip } from "../../components/aiServiceDetail/aiServiceDe
 import { AiServiceDescription } from "../../components/aiServiceDetail/aiServiceDescription/AiServiceDescription";
 
 function AiServiceDetail() {
+  const [introContent, setIntroContent] = useState([]);
+
+  useEffect(() => {
+    const introContentData = [
+      {
+        id: 1,
+        title: "Chat GPT",
+        content: "chan AI가 개발했지요?",
+        url: "https://chat.openai.com/",
+        company: "Open AI",
+        applier: "admin",
+        keyword: ["챗봇", "과제"],
+        popular_job: ["개발자", "디자이너"], // 추가
+        thumbnail:
+          "https://www.headmind.com/wp-content/uploads/2023/01/CHAT-GPT.png",
+        is_liked: false,
+        like_cnt: 599,
+        view_cnt: 1000,
+        rating_point: 3.7,
+        my_rating_point: 5,
+        rating_cnt: 202,
+        views: 402333
+      }
+    ];
+    setIntroContent(introContentData);
+  }, []);
+
   // 탭 기능 구현
   const [currentTab, setCurrentTab] = useState(0);
 
   const tabContents = [
     <AiServiceDescription />,
-    <AiServiceDetailReview />,
+    <AiServiceDetailReview introContent={introContent} />,
     <AiServiceDetailTip />
   ];
 
@@ -23,7 +50,7 @@ function AiServiceDetail() {
 
   return (
     <>
-      <AiServiceDetailIntro />
+      <AiServiceDetailIntro introContent={introContent} />
 
       <S.AiServiceDetailCommentWrap>
         <S.AiServiceDetailCommentCategory>
